+++ CHANGELOG.md
@@ -144,3 +144,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Decode human readable in base 64
 - Log bridgeheads at start
 - Replace hypen in bridgeheads
+- Log bridgeheads sorted